Refactor normalization logic out to util
The normalization logic has been refactored out to util/opimpl.py
